requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
Django==3.2.6
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
